+++ src/Kestrel.Core/Internal/HttpConnectionBuilderExtensions.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Microsoft.AspNetCore.Hosting.Server;
+using Microsoft.AspNetCore.Protocols;
+using Microsoft.AspNetCore.Server.Kestrel.Core.Adapter.Internal;
+
+namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal
+{
+    public static class HttpConnectionBuilderExtensions
+    {
+        public static IConnectionBuilder UseHttpServer<TContext>(this IConnectionBuilder builder, ServiceContext serviceContext, IHttpApplication<TContext> application)
+        {
+            return builder.UseHttpServer(Array.Empty<IConnectionAdapter>(), serviceContext, application);
+        }
+
+        public static IConnectionBuilder UseHttpServer<TContext>(this IConnectionBuilder builder, IList<IConnectionAdapter> adapters, ServiceContext serviceContext, IHttpApplication<TContext> application)
+        {
+            var middleware = new HttpConnectionMiddleware<TContext>(adapters, serviceContext, application);
+            return builder.Use(next =>
+            {
+                return middleware.OnConnectionAsync;
+            });
+        }
+    }
+}
